+++ test/index.test.js
@@ -1,100 +1,6 @@
 import fetchMock from 'fetch-mock';
-import {
-  bestfetch,
-  getRequestKey,
-  duplicateRequests,
-  responseCache,
-  CacheMissError,
-} from '../src';
+import { bestfetch, responseCache, CacheMissError } from '../src';
 import { successfulResponse } from './responses';
-
-describe('duplicateRequests.isRequestInFlight', () => {
-  test('renders false when it is not in flight', () => {
-    expect(duplicateRequests.isRequestInFlight('pasta')).toBe(false);
-  });
-
-  test('renders true when it is not in flight', () => {
-    bestfetch('/test/hangs', { requestKey: 'pasta' });
-    expect(duplicateRequests.isRequestInFlight('pasta')).toBe(true);
-  });
-});
-
-describe('getRequestKey', () => {
-  test('it returns a string, even with no inputs', () => {
-    const key = getRequestKey();
-    expect(typeof key).toBe('string');
-  });
-
-  test('it returns a string, even with an empty object input', () => {
-    const key = getRequestKey({});
-    expect(typeof key).toBe('string');
-  });
-
-  test('it returns a string', () => {
-    const key = getRequestKey({
-      url: '/test/2',
-      method: 'GET',
-    });
-    expect(typeof key).toBe('string');
-  });
-
-  test('it ignores the casing of method', () => {
-    const keyOne = getRequestKey({
-      url: '/test/2',
-      method: 'GET',
-    });
-
-    const keyTwo = getRequestKey({
-      url: '/test/2',
-      method: 'get',
-    });
-
-    expect(keyOne).toBe(keyTwo);
-  });
-
-  test('it produces different keys for different methods', () => {
-    const keyOne = getRequestKey({
-      url: '/test/2',
-      method: 'GET',
-    });
-
-    const keyTwo = getRequestKey({
-      url: '/test/2',
-      method: 'DELETE',
-    });
-
-    expect(keyOne).not.toBe(keyTwo);
-  });
-
-  test('it produces different keys for different URLs', () => {
-    const keyOne = getRequestKey({
-      url: '/test/1',
-      method: 'GET',
-    });
-
-    const keyTwo = getRequestKey({
-      url: '/test/2',
-      method: 'GET',
-    });
-
-    expect(keyOne).not.toBe(keyTwo);
-  });
-
-  test('it produces different keys for different options', () => {
-    const keyOne = getRequestKey({
-      url: '/test/1',
-      method: 'GET',
-    });
-
-    const keyTwo = getRequestKey({
-      url: '/test/1',
-      method: 'GET',
-      body: 'sandwiches',
-    });
-
-    expect(keyOne).not.toBe(keyTwo);
-  });
-});
 
 describe('bestfetch', () => {
   test('only calls fetch once for duplicate requests', () => {
@@ -400,139 +306,6 @@ describe('bestfetch', () => {
         done();
       }
     );
-  });
-});
-
-describe('cacheReadPolicy', () => {
-  test('It errors if you pass an invalid function', () => {
-    expect(() => {
-      responseCache.configureCacheReadPolicy({});
-    }).toThrow();
-  });
-
-  test('Overriding it to ignore the entire cache should work', done => {
-    responseCache.configureCacheReadPolicy(() => false);
-
-    bestfetch('/test/succeeds/json').then(res => {
-      expect(res).toEqual(
-        expect.objectContaining({
-          data: {
-            a: true,
-          },
-          status: 200,
-          statusText: 'OK',
-          ok: true,
-        })
-      );
-
-      bestfetch('/test/succeeds/json').then(resTwo => {
-        expect(resTwo).toEqual(
-          expect.objectContaining({
-            data: {
-              a: true,
-            },
-            status: 200,
-            statusText: 'OK',
-            ok: true,
-          })
-        );
-        expect(fetchMock.calls('/test/succeeds/json').length).toBe(2);
-        done();
-      });
-    });
-  });
-
-  test('Overriding it to ignore cached responses after 1 read should work', done => {
-    responseCache.configureCacheReadPolicy(cacheObject => {
-      return cacheObject.accessCount < 1;
-    });
-
-    bestfetch('/test/succeeds/json').then(res => {
-      expect(res).toEqual(
-        expect.objectContaining({
-          data: {
-            a: true,
-          },
-          status: 200,
-          statusText: 'OK',
-          ok: true,
-        })
-      );
-
-      bestfetch('/test/succeeds/json').then(resTwo => {
-        expect(resTwo).toEqual(
-          expect.objectContaining({
-            data: {
-              a: true,
-            },
-            status: 200,
-            statusText: 'OK',
-            ok: true,
-          })
-        );
-        expect(fetchMock.calls('/test/succeeds/json').length).toBe(1);
-
-        bestfetch('/test/succeeds/json').then(resThree => {
-          expect(resThree).toEqual(
-            expect.objectContaining({
-              data: {
-                a: true,
-              },
-              status: 200,
-              statusText: 'OK',
-              ok: true,
-            })
-          );
-          expect(fetchMock.calls('/test/succeeds/json').length).toBe(2);
-          done();
-        });
-      });
-    });
-  });
-});
-
-describe('cacheWritePolicy', () => {
-  test('It errors if you pass an invalid function', () => {
-    expect(() => {
-      responseCache.configureCacheWritePolicy({});
-    }).toThrow();
-  });
-
-  test('default does not cache 500 server errors', done => {
-    bestfetch('/test/fails/internal-server-error', {
-      requestKey: 'will-error',
-      responseType: 'text',
-    }).then(res => {
-      expect(res).toEqual(
-        expect.objectContaining({
-          data: 'Server error message',
-          status: 500,
-          statusText: 'Internal Server Error',
-          ok: false,
-        })
-      );
-
-      expect(responseCache.has('will-error')).toBe(false);
-
-      bestfetch('/test/fails/internal-server-error', {
-        requestKey: 'will-error',
-        responseType: 'text',
-      }).then(resTwo => {
-        expect(resTwo).toEqual(
-          expect.objectContaining({
-            data: 'Server error message',
-            status: 500,
-            statusText: 'Internal Server Error',
-            ok: false,
-          })
-        );
-        expect(responseCache.has('will-error')).toBe(false);
-        expect(
-          fetchMock.calls('/test/fails/internal-server-error').length
-        ).toBe(2);
-        done();
-      });
-    });
   });
 });
